==> 02a_CTSN_RNAseq.rmd <==
---
title: "CTSN_RNAseq"
author: "Francis Leblanc"
date: '2023-01-24'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = TRUE,
  warning = TRUE
)

library(tximport)
library(DESeq2)
library(factoextra)
library(PCAForQTL)
library(dplyr)
set.seed(2023)
```

# CTSN bulk 

```{r}
tx2gene <- rio::import("data/external/tx2gene.gencode.v32.tsv.gz")
all_meta <- rio::import("data/CTSN/RNAseq/meta.ctsn.tsv") %>%
  dplyr::rename(sex = SEX) %>%
  mutate(path = paste0("data/stranded/",
                       sample,
                       "_/abundance.h5"))

txi.kallisto <- tximport(all_meta$path,
                         type = "kallisto",
                         tx2gene = tx2gene[, 1:2],
                         ignoreAfterBar = T)

dds <- DESeqDataSetFromTximport(txi.kallisto,
                                all_meta,
                                design = ~ sex + condition)

counts <- dds@assays@data$counts
colnames(counts) <- all_meta$sample
saveRDS(counts, "data/CTSN/RNAseq/counts.rds")

meta.ctsn <- dplyr::select(all_meta, sample, sex, condition)
saveRDS(meta.ctsn, "data/CTSN/RNAseq/meta.ctsn.rds")
```

## output vst expression

```{r}
tx2gene <- rio::import("data/external/tx2gene.gencode.v32.tsv.gz")
counts <- readRDS("data/CTSN/RNAseq/counts.rds")
rio::export(as.data.frame(counts),
            "data/CTSN/RNAseq/counts.ctsn.tsv",
            row.names = T)

meta <- readRDS("data/CTSN/RNAseq/meta.ctsn.rds")
row.names(meta) <- meta$sample
rio::export(meta, "data/CTSN/RNAseq/meta.ctsn.tsv")

dds <- DESeqDataSetFromMatrix(counts, meta, design = ~ sex + condition)
vst <- vst(dds, blind = T)

vst.df <- assay(vst)
saveRDS(vst.df, "data/CTSN/RNAseq/vst.rds")
```

## output DEG
```{r}
dds$condition <- relevel(dds$condition, ref = "SR")
min_counts <- rowSums(counts(dds)) >= 10
dds <- dds[min_counts,]

dds <- DESeq(dds)
res <- lfcShrink(dds, coef = "condition_AF_vs_SR", type = "ashr")

res <- as.data.frame(res)
res$gene_name <- tx2gene$SYMB[match(row.names(res), tx2gene$GENEID)]
saveRDS(res, "data/CTSN/RNAseq/DEG.rds")
```


## PCA Qc
```{r}
plotPCA(vst, intgroup = "sex", ntop = 500, returnData = FALSE)
ggsave("figs/CTSN_RNAseq/PCA.png")
```


### Test k PCs as covariates

```{r}
rv <- rowVars(as.matrix(vst.df))
select <- rv > 0.001
pca <- prcomp(t(vst.df[select,]), center = TRUE, scale. = TRUE)

# determine number of PCs to keep as covariates in eQTL
resultRunElbow <- runElbow(prcompResult = pca)
resultRunBE <- runBE(t(vst.df[select,]),
                     B = 20,
                     alpha = 0.05,
                     mc.cores = 1,
                     verbose = T)

K_elbow <- resultRunElbow # 7
K_BE <- resultRunBE$numOfPCsChosen # 13

# GTEx uses 60 PEER factors, and they are almost identical to the top 60 PCs.
K_GTEx <- 60

png("figs/CTSN_RNAseq/ScreePlot_CTSN_RNAseq.PCs.png",
    width = 7,
    height = 5,
    res = 300,
    units = "in")

PCAForQTL::makeScreePlot(pca,
                         labels = c("Elbow", "BE", "GTEx"),
                         values = c(K_elbow, K_BE, K_GTEx),
                         titleText = "CTSN RNAseq")
dev.off()

covariatesToUse <- cbind(meta[, c("sex", "condition")], pca$x[, 1:7])
saveRDS(covariatesToUse, "data/CTSN/RNAseq/meta_cov.7PCs.rds")
```
