==> 09_Demo.rmd <==
---
title: "Demo.fisher"
author: "Francis Leblanc"
date: "2023-06-30"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
```

```{r}
demo <- rio::import("results/TableS1_noP.xlsx")
colnames(demo) <- demo[2,]
demo <- demo[-c(1:3),]

demo$CTSN_Fisher_P <- lapply(1:nrow(demo), function(x){
  cases <- as.numeric(demo[x,2:3])
  if (!any(is.na(cases))) {
    df <- rbind(cases, 31 - cases)
    fisher.test(df)$p.value
  } else(return(NA))
  
}) %>% unlist()

demo$Harbin_Fisher_P <- lapply(1:nrow(demo), function(x){
  cases <- as.numeric(demo[x,4:5])
  if (!any(is.na(cases))) {
    df <- rbind(cases, c(28 - cases[1], 37 - cases[2]))
    fisher.test(df)$p.value
  } else(return(NA))
  
}) %>% unlist()


# harbin age t-test
age.h <- rio::import("data/Harbin/age.csv") 
age.h$rhythm <- limma::strsplit2(age.h$Patient,split = "-")[,1]
t.test(as.numeric(age.h$Age) ~ age.h$rhythm)
# 	Welch Two Sample t-test
# 
# data:  as.numeric(age.h$Age) by age.h$rhythm
# t = 0.49226, df = 62.949, p-value = 0.6242
# alternative hypothesis: true difference in means between group AF and group SR is not equal to 0
# 95 percent confidence interval:
#  -3.245628  5.367249
# sample estimates:
# mean in group AF mean in group SR 
#         59.25000         58.18919 
        
# CTSN age t-test
meta <- readRDS("data/CTSN/RNAseq/meta.ctsn.rds")
meta$age <- rio::import("data/CTSN/age.csv")$age
t.test(as.numeric(meta$age) ~ meta$condition)
# 	Welch Two Sample t-test
# 
# data:  as.numeric(meta$age) by meta$condition
# t = 1.8135, df = 59.783, p-value = 0.07477
# alternative hypothesis: true difference in means between group AF and group SR is not equal to 0
# 95 percent confidence interval:
#  -0.3757367  7.6660592
# sample estimates:
# mean in group AF mean in group SR 
#         69.25806         65.61290

rio::export(demo, "results/TableS1_P.xlsx")
```
